my_app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
pytest
pytest-django